backend/requirements.txt
django>=3.1.12
psycopg2-binary==2.8.5
djangorestframework>=3.11.2
django-model-utils==4.1.1
django-cors-headers==3.10.0
pytz==2021.3
python-dateutil==2.8.2